chatbot/requirements.txt
langgraph
langsmith
langchain
langchain-text-splitters
langchain[openai]
IPython
langchain-tavily
langchain-community
faiss-cpu
sentence-transformers
flask
flask-cors
Pillow
rapidfuzz
pymongo
mistralai
python-dotenv